fix(dialog): moved CreateUserDialog.vue in correct folder
